First commit to the project
This is the core of a small application that will show how to upload and
download files from S3 in Go. This, the first change, includes the
minimum code required to provide a working web application; because web
applications are the kind of code that I most like to write.